requirements.txt
# Quiz Generator Dependencies
langchain>=1.0.0
langchain-core>=1.0.0
langchain-openai>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
